requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
scikit-learn
pydantic
python-dotenv
openai
